refactor: extract lookupPerson query and deletePerson service from action
Move 200-line Prisma findFirst into getPersonWithAllRelations() in query
layer and cascade delete transaction into deletePerson() in service layer.
Actions become thin orchestrators.
